Remove INSTALL_NAME_DIR definition
Defining INSTALL_NAME_DIR CMake variable forces the install name to be an
absolute path instead of `@rpath/{target_name}` [1]. This causes problems
when one tries to relocate the target once it is installed.
This patch removes this target property. If a project needs to keep
the absolute path, INSTALL_NAME_DIR can be set during project configuration.

[1] https://gitlab.kitware.com/cmake/cmake/issues/16589